libcalib: split into calibration `libcalib` and correction `libcorr`
 - split done to avoid circular dependencies

JIRA: PP-89
