+++ libs/libcalib/hmap.h
