+++ libs/libcalib/libcalib.h
@@ -31,27 +31,34 @@
 typedef struct {
 	char name[16]; /* alias of this calibration */
 
-	/* process related */
-	int (*run)(void);                /* calibration procedure start */
-	int (*done)(void);               /* procedure deinitialization */
-	int (*init)(int, const char **); /* procedure initialization */
+	union {
+		struct {
+			int (*run)(void);                /* calibration procedure start */
+			int (*done)(void);               /* procedure deinitialization */
+			int (*init)(int, const char **); /* procedure initialization */
+		} calib;
+
+		struct {
+			/* Correction calculation prodecures */
+			int (*perform)(sensor_event_t *); /* corrects given measurement event based on own correction type */
+			int (*init)(void);                /* initialization of correction algorithm, NULL if unnecessary */
+			int (*done)(void);                /* deinitialization of correction algorithm, NULL if unnecessary */
+			int (*recalc)(void);              /* correction recalculation procedure */
+			const time_t delay;               /* time delay in microseconds between correction recalculation, 0 if correction is time-invariant */
+		} corr;
+	} proc;
 
 	/* utility related */
 	const char *(*help)(void);             /* help message description */
 	int (*interpret)(const char *, float); /* calibration file data interpreter */
 	int (*write)(FILE *);                  /* calibration file data write */
 
-	/* Correction calculation prodecures */
-	int (*corrDo)(sensor_event_t *); /* corrects given measurement event based on own correction type */
-	int (*corrInit)(void);           /* initialization of correction algorithm, NULL if unnecessary */
-	int (*corrDone)(void);           /* deinitialization of correction algorithm, NULL if unnecessary */
-	int (*corrRecalc)(void);         /* correction recalculation procedure */
-	const time_t delay;              /* time delay in microseconds between correction recalculation, 0 if correction is time-invariant */
 } calib_t;
 
 
-/* registering new calibration procedure */
-extern void calib_register(calib_t *c);
+/* TODO: common_register needs to be moved to some internal header file */
+/* registering new calibration/correction procedure */
+extern void common_register(calib_t *c); 
 
 /*
 * Reads calibration paramaeters file and writes its data to procedures in `hm`
